Auth implemented

+++ src/accounts/api/v1/Accounts_blueprint.py
@@ -4,16 +4,21 @@ from quart import Blueprint, request, abort, jsonify
 from quart_schema import validate_request, validate_response, tag, document_request, document_response, \
     ResponseSchemaValidationError
 
-from ...models.Accounts import AccountCreate, AccountUpdate, AccountView, AccountUpdateBalance, AccountListResponse
+from ...models.Accounts import AccountCreate, AccountUpdate, AccountView, AccountUpdateBalance, AccountListResponse, \
+    AccountAuthCreate
+from ...models.Auth import CreaterAuthUser
 from ...models.Cards import DeleteCardRequest
 from ...models.Empty import EmptyGet404, EmptyPatch400, EmptyPatch403, EmptyPatch404, EmptyPost400, \
     EmptyPost404, EmptyError503, EmptyDelete204, EmptyGet400, EmptyDelete400, EmptyDelete404, EmptyDelete200, \
-    EmptyPatch204
+    EmptyPatch204, EmptyPost403
 
 from ...services.Accounts_service import AccountService
+from ...services import Auth_service
 
 from logging import getLogger
 from ...core.config import settings
+
+import jwt
 
 logger = getLogger()
 logger.setLevel(settings.LOG_LEVEL)
@@ -21,15 +26,26 @@ logger.setLevel(settings.LOG_LEVEL)
 bp = Blueprint("accounts_v1", __name__, url_prefix="/v1/accounts")
 
 @bp.post("/")
-@validate_request(AccountCreate)
+@validate_request(AccountAuthCreate)
 @validate_response(AccountView, 201)
 @document_response(EmptyPost400, 400)
 @tag(["v1"])
-async def create_account(data: AccountCreate):
+async def create_account(data: AccountAuthCreate):
     service = AccountService()
-    res = await service.create_new_account(data)
+    data_account = AccountCreate(name=data.name, email=data.email, subscription=data.subscription)
+    res = await service.create_new_account(data_account)
     if isinstance(res, EmptyPost404):
         abort(400, description="Bad Request")
+    
+    # TODO AÑADIR EL IBAN, NO PUEDE CREARLO AUTH
+    auth_create = CreaterAuthUser(email=res.email, name=res.name, password=data.password, phoneNumber=data.phoneNumber)
+    res2 = await Auth_service.create_user(auth_create)
+    
+    
+    if isinstance(res2, EmptyError503):
+        await service.delete_account(res.iban)
+        abort(503, description="Auth microservice is unavailable")
+    
     return res, 201
 
 @bp.get("/<string:iban>")
@@ -38,6 +54,15 @@ async def create_account(data: AccountCreate):
 @document_response(EmptyGet404, 404)
 @tag(["v1"])
 async def view_account(iban: str):
+    auth_header = request.headers.get('Authorization')
+    if not auth_header:
+        return jsonify({"error": "Falta el header Authorization"}), 401
+    _, token = auth_header.split(" ")
+    jwt_data = decode_jwt(token)
+    jwt_iban = jwt_data.get('iban')
+    if jwt_iban != iban:
+        abort(403, description="Unauthorized access")
+    
     service = AccountService()
     res = await service.get_account_by_iban(iban)
     if isinstance(res, EmptyGet404):
@@ -67,6 +92,15 @@ async def view_accounts():
 @document_response(EmptyPatch404, 404)
 @tag(["v1"])
 async def update_account(iban: str):
+    auth_header = request.headers.get('Authorization')
+    if not auth_header:
+        return jsonify({"error": "Falta el header Authorization"}), 401
+    _, token = auth_header.split(" ")
+    jwt_data = decode_jwt(token)
+    jwt_iban = jwt_data.get('iban')
+    if jwt_iban != iban:
+        abort(403, description="Unauthorized access")
+    
     service = AccountService()
     
     # Validate JSON here
@@ -86,6 +120,15 @@ async def update_account(iban: str):
 @document_response(EmptyPatch404, 404)
 @tag(["v1"])
 async def update_account_balance(iban: str, currency: str):
+    auth_header = request.headers.get('Authorization')
+    if not auth_header:
+        return jsonify({"error": "Falta el header Authorization"}), 401
+    _, token = auth_header.split(" ")
+    jwt_data = decode_jwt(token)
+    jwt_iban = jwt_data.get('iban')
+    if jwt_iban != iban:
+        abort(403, description="Unauthorized access")
+    
     service = AccountService()
     
     # Validate JSON here
@@ -111,6 +154,15 @@ async def update_account_balance(iban: str, currency: str):
 @document_response(EmptyDelete404, 404)
 @tag(["v1"])
 async def delete_account(iban: str):
+    auth_header = request.headers.get('Authorization')
+    if not auth_header:
+        return jsonify({"error": "Falta el header Authorization"}), 401
+    _, token = auth_header.split(" ")
+    jwt_data = decode_jwt(token)
+    jwt_iban = jwt_data.get('iban')
+    if jwt_iban != iban:
+        abort(403, description="Unauthorized access")
+    
     service = AccountService()
     res = await service.delete_account(iban)
     
@@ -127,6 +179,15 @@ async def delete_account(iban: str):
 @document_response(EmptyPatch404, 404)
 @tag(["v1"])
 async def block_account(iban: str):
+    auth_header = request.headers.get('Authorization')
+    if not auth_header:
+        return jsonify({"error": "Falta el header Authorization"}), 401
+    _, token = auth_header.split(" ")
+    jwt_data = decode_jwt(token)
+    jwt_iban = jwt_data.get('iban')
+    if jwt_iban != iban:
+        abort(403, description="Unauthorized access")
+        
     service = AccountService()
     res = await service.block_account(iban)
     if isinstance(res, EmptyPatch400):
@@ -142,6 +203,15 @@ async def block_account(iban: str):
 @document_response(EmptyPatch404, 404)
 @tag(["v1"])
 async def unblock_account(iban: str):
+    auth_header = request.headers.get('Authorization')
+    if not auth_header:
+        return jsonify({"error": "Falta el header Authorization"}), 401
+    _, token = auth_header.split(" ")
+    jwt_data = decode_jwt(token)
+    jwt_iban = jwt_data.get('iban')
+    if jwt_iban != iban:
+        abort(403, description="Unauthorized access")
+        
     service = AccountService()
     res = await service.unblock_account(iban)
     if isinstance(res, EmptyPatch400):
@@ -154,15 +224,27 @@ async def unblock_account(iban: str):
 @validate_response(AccountView, 200)
 @document_response(EmptyPost400, 400)
 @document_response(EmptyPost404, 404)
+@document_response(EmptyPost403, 403)
 @document_response(EmptyError503, 503)
 @tag(["v1"])
 async def create_card_account(iban: str):
+    auth_header = request.headers.get('Authorization')
+    if not auth_header:
+        return jsonify({"error": "Falta el header Authorization"}), 401
+    _, token = auth_header.split(" ")
+    jwt_data = decode_jwt(token)
+    jwt_iban = jwt_data.get('iban')
+    if jwt_iban != iban:
+        abort(403, description="Unauthorized access")
+    
     service = AccountService()
     res = await service.account_create_card(iban)
     if isinstance(res, EmptyPost400):
         abort(400, description="Bad Request")
     if isinstance(res, EmptyPost404):
         abort(404, description="Account not found")
+    if isinstance(res, EmptyPost403):
+        abort(403, description="Cannot create more cards with your subscription. Consider upgrading.")
     if isinstance(res, EmptyError503):
         abort(503, description="Microservice cards is unavailable")
     return res
@@ -175,6 +257,15 @@ async def create_card_account(iban: str):
 @document_response(EmptyError503, 503)
 @tag(["v1"])
 async def delete_card_account(iban: str, data: DeleteCardRequest):
+    auth_header = request.headers.get('Authorization')
+    if not auth_header:
+        return jsonify({"error": "Falta el header Authorization"}), 401
+    _, token = auth_header.split(" ")
+    jwt_data = decode_jwt(token)
+    jwt_iban = jwt_data.get('iban')
+    if jwt_iban != iban:
+        abort(403, description="Unauthorized access")
+    
     service = AccountService()
     res = await service.account_delete_card(iban, data)
     if isinstance(res, EmptyPost400):
@@ -192,3 +283,6 @@ async def handle_validation_error(error):
         "message": "Data input not valid",
         "details": error.validation_error.errors()
     }), 400
+
+def decode_jwt(token):
+    return jwt.decode(token, options={"verify_signature": False})
